Reading Gaia Files
We now have a reading of Gaia files.
The gaia file in the Gifs folder has contents

```
gaia
tag: Use the force, Luke
```

And we have

$ stack exec gaia-utils run-query "Force"
/Users/pascal/Desktop
    /Users/pascal/Desktop/Gifs

ie: case insensitive parsing of the tag body causing the parent
to be returned.

Also I have renamed the Types of the Gaia Files Directive parsing
because they conflict with all the Nyx stuff that is coming...